Develop

+++ app/Http/Controllers/BranchController.php
@@ -13,6 +13,7 @@ use App\Tag;
 use App\TagBranch;
 use Validator;
 use JWTAuth;
+use App\UserInvitation;
 
 class BranchController extends Controller
 {
@@ -74,6 +75,14 @@ class BranchController extends Controller
 			->get();
 
 			$branch->tags = $tags;
+
+			$verifications = \DB::table('branch_verification AS bv')
+			->where('bv.branch_id','=',$branch->id)
+			->select('bv.verification_type','bv.description','bv.id')
+			->get();
+
+			$branch->verifications = $verifications;
+
 		}
 
 		$response = ['code' => 200,'count' => $count,'data' => $branches];
@@ -114,57 +123,69 @@ class BranchController extends Controller
 
 		//SE OBTIENE EL ID DE LA COMPANY QUE LE PERTENCE LA BRANCH
 		$company_id = $request->company_id;
-		$company = Company::find($company_id);
+		$company = Company::with('branches')->find($company_id);
 
 		if(!is_null($company)){
 
 			$userRequested = \Auth::User();
 			$role = $userRequested->authRole;
 
+			//SE VALIDA QUE EL USUARIO NO TENGA BRANCHES O TENGA HABILITADO LA CREACION DE MULTIPLES COMPAÑIAS/SUCURSALES
+			//PARA PODER GUARDAR UNA NUEVA
+			if(($userRequested->enabled_companies == \Config::get('app.NO_ENABLED_COMPANIES') && 
+				$company->branches->count() == 0) || $userRequested->enabled_companies == \Config::get('app.ENABLED_COMPANIES')){
 
-			//SE VERIFICA QUE EL USER QUE HIZO LA PETICION SOLO PUEDA GUARDAR BRANCHES EN SUS COMPANIES
-			if(($userRequested->id == $company->user_id) || $role == 'ADMIN'){
+				//SE VERIFICA QUE EL USER QUE HIZO LA PETICION SOLO PUEDA GUARDAR BRANCHES EN SUS COMPANIES
+				if(($userRequested->id == $company->user_id) || $role == 'ADMIN'){	
 
-				//SE UNA INSTANCIA DE BRANCH
-				$branch = new Branch;
-				$branch->company_id = $company_id;
-				$branch->address = $request->address;
-				$branch->phone = $request->phone;
-				$branch->latitude = $request->latitude;
-				$branch->longitude = $request->longitude;
-				$branch->state_id = $request->state_id;
-				$branch->schedule = $request->schedule;
-				$branch->name = $request->name;
-                $branch->geom = [$request->longitude, $request->latitude];
+					//Si se detecta un codigo de invitacion entonces de borra el codigo
+					if($request->code)
+						$this::clearCode($request->code,$company->user_id);
+					
+					//SE UNA INSTANCIA DE BRANCH
+					$branch = new Branch;
+					$branch->company_id = $company_id;
+					$branch->address = $request->address;
+					$branch->phone = $request->phone;
+					$branch->latitude = $request->latitude;
+					$branch->longitude = $request->longitude;
+					$branch->state_id = $request->state_id;
+					$branch->schedule = $request->schedule;
+					$branch->name = $request->name;
+					$branch->geom = [$request->longitude, $request->latitude];
 
-				$branch->save();
+					$branch->save();
 
-				//SE GUARDAN LOS TAGS QUE YA EXISTEN EN LA DB EN LA BRANCH
-				$this->saveTag($request->tag,$branch);
+					//SE GUARDAN LOS TAGS QUE YA EXISTEN EN LA DB EN LA BRANCH
+					$this->saveTag($request->tag,$branch);
 
-				//SE GUARDAN LOS NUEVOS TAGS CREADOS POR EL USER
-				$this->newTag($request->tag_new,$company->category_id,$branch);
+					//SE GUARDAN LOS NUEVOS TAGS CREADOS POR EL USER
+					$this->newTag($request->tag_new,$company->category_id,$branch);
 
-				//SE VALIDA QUE LA BRANCH SE HALLA GUARDADO
-				if($branch != false){
+					//SE VALIDA QUE LA BRANCH SE HALLA GUARDADO
+					if($branch != false){
 
-					//SE OBTINEN TODOS LOS TAGS DE LA BRANCH CREADA PARA UNIRLA AL JSON
-					$tags = \DB::table('tags_branches')
-					->join('tags','tags_branches.tag_id','=','tags.id')
-					->where('branch_id','=',$branch->id)
-					->select('tags.id','tags.name','tags.description')
-					->get();
+						//SE OBTINEN TODOS LOS TAGS DE LA BRANCH CREADA PARA UNIRLA AL JSON
+						$tags = \DB::table('tags_branches')
+						->join('tags','tags_branches.tag_id','=','tags.id')
+						->where('branch_id','=',$branch->id)
+						->select('tags.id','tags.name','tags.description')
+						->get();
 
-					$branch->tags = $tags;
+						$branch->tags = $tags;
 
-					$response = ['data' => $branch,'code' => 200,'message' => 'Branch was created succefully'];
-					return response()->json($response,200);
+						$response = ['data' => $branch,'code' => 200,'message' => 'Branch was created succefully'];
+						return response()->json($response,200);
+					}else{
+						$response = ['error' => 'It has occurred an error trying to save the branch','code' => 500];
+						return response()->json($response,500);
+					}
 				}else{
-					$response = ['error' => 'It has occurred an error trying to save the branch','code' => 500];
-					return response()->json($response,500);
+					$response = ['error'   => 'Unauthorized2','code' => 403];
+					return response()->json($response, 403);
 				}
 			}else{
-				$response = ['error'   => 'Unauthorized','code' => 403];
+				$response = ['error'   => 'Unauthorized1','code' => 403];
 				return response()->json($response, 403);
 			}
 
@@ -175,6 +196,17 @@ class BranchController extends Controller
 		}
 
     }
+	
+	/*
+	* Borra el código de invitación utilizado para guardar una branch y
+	* actualiza el id del usuario que usó el código
+	*/
+	private function clearCode($code,$user_id){
+		$invitation = UserInvitation::where('code','=',$code)->get()->first();
+		$invitation->to_user_id = $user_id;
+		$invitation->save();
+		$invitation->delete();
+	}
 
     /**
      * Display the specified resource.
@@ -188,6 +220,7 @@ class BranchController extends Controller
         $branch = Branch::with('company')
                         ->with('state')
                         ->with('state.country')
+						->with('verifications')
                         ->where('id','=',$id)
                         ->first();
 
@@ -243,8 +276,8 @@ class BranchController extends Controller
 
 		//SE VERIFICA SI ALGUN CAMPO NO ESTA CORRECTO
 		if($v->fails()){
-			$response = ['error' => 'Bad Request', 'data' => $v->messages(), 'code' =>  422];
-			return response()->json($response,422);
+			$response = ['error' => 'Bad Request', 'data' => $v->messages(), 'code' =>  400];
+			return response()->json($response, 400);
 		}
 
 		//SE OBTIENE LA BRANCH SOLICITIDA JUNTO CON LA COMPANY QUE LE PERTENECE
@@ -268,6 +301,7 @@ class BranchController extends Controller
 				$branch->schedule = $request->schedule;
 				$branch->name = $request->name;
 				$branch->role_id = $userRequested->id;
+                $branch->geom = [$request->longitude, $request->latitude];
 				$branch->role = $this->user_roles[$userRequested->roleAuth];
 
 				$branch->save();
@@ -306,8 +340,8 @@ class BranchController extends Controller
 
 		}else{
 			//EN DADO CASO QUE EL ID DE BRANCH NO SE HALLA ENCONTRADO
-			$response = ['error' => 'Branch does not exist','code' => 422];
-			return response()->json($response,422);
+			$response = ['error' => 'Branch does not exist','code' => 404];
+			return response()->json($response, 404);
 		}
 
     }
@@ -331,27 +365,35 @@ class BranchController extends Controller
 			//SE OBTIENE LA COMPANY DE LA BRANCH
 			$company = $branch->company;
 
-			//SE VERIFICA QUE EL USER QUE HIZO LA PETICION SOLO PUEDA ELIMINAR SUS BRANCHES
-			if($userRequested->id == $company->user_id || $userRequested->roleAuth == "ADMIN"){
+			//SE VALIDA QUE EL USUARIO TENGA HABILITADO LA CREACION DE MULTIPLES COMPAÑIAS/SUCURSALES
+			//PARA PODER ELIMINAR UNA SUCURSAL
+			if($userRequested->enabled_companies == \Config::get('app.ENABLED_COMPANIES')){
 
-				$branch->role_id = $userRequested->id;
-				$branch->role = $this->user_roles[$userRequested->roleAuth];
-				$branch->save();
+				//SE VERIFICA QUE EL USER QUE HIZO LA PETICION SOLO PUEDA ELIMINAR SUS BRANCHES
+				if($userRequested->id == $company->user_id || $userRequested->roleAuth == "ADMIN"){
 
-				//SE BORRAR LA BRANCH
-				$rows = $branch->delete();
+					$branch->role_id = $userRequested->id;
+					$branch->role = $this->user_roles[$userRequested->roleAuth];
+					$branch->save();
 
-				//SE ELIMINAN TODAS LAS TAG QUE LE PERTENECEN A LA BRANCH
-				/*$timestamp = time()+date('Z');
-				$date = date('Y-m-d H:i:s',$timestamp);
-				\DB::update("UPDATE FROM tags_branches SET deteted_at = NOW WHERE branch_id = ".$branch->id." ");	*/
+					//SE BORRAR LA BRANCH
+					$rows = $branch->delete();
 
-				if($rows > 0){
-					$response = ['code' => 200,'message' => "Branch was deleted succefully"];
-					return response()->json($response,200);
+					//SE ELIMINAN TODAS LAS TAG QUE LE PERTENECEN A LA BRANCH
+					/*$timestamp = time()+date('Z');
+					$date = date('Y-m-d H:i:s',$timestamp);
+					\DB::update("UPDATE FROM tags_branches SET deteted_at = NOW WHERE branch_id = ".$branch->id." ");	*/
+
+					if($rows > 0){
+						$response = ['code' => 200,'message' => "Branch was deleted succefully"];
+						return response()->json($response,200);
+					}else{
+						$response = ['error' => 'It has occurred an error trying to delete the branch','code' => 500];
+						return response()->json($response,500);
+					}
 				}else{
-					$response = ['error' => 'It has occurred an error trying to delete the branch','code' => 500];
-					return response()->json($response,500);
+					$response = ['error'   => 'Unauthorized','code' => 403];
+					return response()->json($response, 403);
 				}
 			}else{
 				$response = ['error'   => 'Unauthorized','code' => 403];
@@ -383,20 +425,25 @@ class BranchController extends Controller
 
 				for($i = 0;$i < count($tags);$i++){
 					$tag = (object) $tags[$i];
+					
+					$tag_verification = Tag::find($tag->tag_id);
+					
+					//Si el tag existe entonces se guarda en la db
+					if($tag_verification){
+						$row = \DB::table('tags_branches')->insert(
+								[
+											'tag_id' => $tag->tag_id,
+											'branch_id' => $branch->id,
+											'created_at' => date('Y-m-d h:i:s',time()),
+											'updated_at' => date('Y-m-d h:i:s',time())
+								]
+							);
 
-					$row = \DB::table('tags_branches')->insert(
-							[
-										'tag_id' => $tag->tag_id,
-										'branch_id' => $branch->id,
-                                        'created_at' => date('Y-m-d h:i:s',time()),
-                                        'updated_at' => date('Y-m-d h:i:s',time())
-							]
-						);
-
-					//SE VALIDA QUE EL TAG SE GUARDO CORRECTAMENTE
-					if($row != true){
-						$response = ['error' => 'It has occurred an error trying to save tags','code' => 500];
-						return response()->json($response,500);
+						//SE VALIDA QUE EL TAG SE GUARDO CORRECTAMENTE
+						if($row != true){
+							$response = ['error' => 'It has occurred an error trying to save tags','code' => 500];
+							return response()->json($response,500);
+						}
 					}
 				}
 			}
